requirements.txt
# Flask Core
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5
Flask-WTF>=1.2.1
Flask-Login>=0.6.3
Flask-Security-Too>=5.3.3
Flask-Admin>=1.6.1
Flask-SocketIO>=5.3.6

# Database
PyMySQL>=1.1.0
cryptography>=41.0.0
redis>=5.0.0
SQLAlchemy>=2.0.23

# Background Jobs
celery>=5.3.4
APScheduler>=3.10.4

# External Integrations (authorizenet removed - incompatible with Python 3.13, will use httpx directly)
sendgrid>=6.11.0
openai>=1.6.0
boto3>=1.34.0
httpx>=0.26.0

# Observability
structlog>=23.2.0
prometheus-flask-exporter>=0.23.0

# Utilities
python-dotenv>=1.0.0
email-validator>=2.1.0
Pillow>=10.2.0
passlib>=1.7.4
bcrypt>=4.1.2

# WebSocket
python-socketio>=5.10.0
gevent>=24.2.1
gevent-websocket>=0.10.1
